+++ src/main/java/com/smooth/driving_analysis_service/progress/entity/DrivingRecordDomain.java
@@ -1,5 +1,5 @@
 // src/main/java/com/smooth/driving_analysis_service/progress/domain/TripSummary.java
-package com.smooth.driving_analysis_service.progress.domain;
+package com.smooth.driving_analysis_service.progress.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
