+++ src/main/java/com/smooth/driving_analysis_service/progress/repository/VUserProgress15Repository.java
@@ -1,6 +1,6 @@
 package com.smooth.driving_analysis_service.progress.repository;
 
-import com.smooth.driving_analysis_service.progress.domain.VUserProgress15Domain;
+import com.smooth.driving_analysis_service.progress.entity.VUserProgress15Domain;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface VUserProgress15Repository extends JpaRepository<VUserProgress15Domain, String> {
