+++ src/main/java/com/smooth/driving_analysis_service/progress/service/EligibilityService.java
@@ -2,7 +2,7 @@ package com.smooth.driving_analysis_service.progress.service;
 
 import com.smooth.driving_analysis_service.progress.dto.EligibilityDto;
 import com.smooth.driving_analysis_service.progress.dto.ProgressDto;
-import com.smooth.driving_analysis_service.progress.domain.VUserProgress15Domain;
+import com.smooth.driving_analysis_service.progress.entity.VUserProgress15Domain;
 import com.smooth.driving_analysis_service.progress.repository.VUserProgress15Repository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
